refactor: 패키지 구조 변경 및 기물별 표현값 조회 기능 개선
[- Game, Board 패키지 추가
- 기존 O(N) 연산으로 기물별 표현값을 조회하는 기능을 Map 기반 조회 기능으로 구현하도록 리팩토링]
