+++ src/main/java/com/wootecam/chess/pieces/Color.java
@@ -1,14 +1,9 @@
 package com.wootecam.chess.pieces;
 
 public enum Color {
-    WHITE("white"),
-    BLACK("black");
-
-    public final String displayName;
-
-    Color(String displayName) {
-        this.displayName = displayName;
-    }
+    WHITE,
+    BLACK,
+    ;
 
     public boolean isWhite() {
         return this == WHITE;
